Made SoundChannel use a darray to store blocks instead of a llist. This means that the block can be found by simply dividing the sample index by the number of samples in a block. This won't work once we support insert and delete into/out-of the sound, but we can do something better then.
Also fixed an int that should have been an int64 (SoundChannel::GetSoundPosFromSampleIdx()). This had been wrong since almost the beginning.

+++ src/sound_channel.cpp
@@ -18,7 +18,7 @@ unsigned SoundChannel::GetLength()
 }
 
 
-SoundChannel::SoundPos SoundChannel::GetSoundPosFromSampleIdx(int sample_idx)
+SoundChannel::SoundPos SoundChannel::GetSoundPosFromSampleIdx(int64_t sample_idx)
 {
     int block_idx = 0;
